Move LineLabelNotUsedInspectionTests for QuickFixes into relevant files under RD.QuickFixes
