+++ RubberduckTests/QuickFixes/IgnoreOnceQuickFixTests.cs
@@ -386,6 +386,32 @@ End Function";
             Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
         }
 
+        [TestMethod]
+
+        [TestCategory("QuickFixes")]
+        public void LabelNotUsed_IgnoreQuickFixWorks()
+        {
+            const string inputCode =
+@"Sub Foo()
+label1:
+End Sub";
+
+            const string expectedCode =
+@"Sub Foo()
+'@Ignore LineLabelNotUsed
+label1:
+End Sub";
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new LineLabelNotUsedInspection(state);
+            new IgnoreOnceQuickFix(state, new[] { inspection }).Fix(inspection.GetInspectionResults().First());
+
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+
 
     }
 }
