+++ RubberduckTests/QuickFixes/RemoveUnusedDeclarationQuickFixTests.cs
@@ -37,5 +37,63 @@ End Sub";
             var rewrittenCode = rewriter.GetText();
             Assert.AreEqual(expectedCode, rewrittenCode);
         }
+
+
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void LabelNotUsed_QuickFixWorks()
+        {
+            const string inputCode =
+@"Sub Foo()
+label1:
+End Sub";
+
+            const string expectedCode =
+@"Sub Foo()
+
+End Sub";
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new LineLabelNotUsedInspection(state);
+            new RemoveUnusedDeclarationQuickFix(state).Fix(inspection.GetInspectionResults().First());
+
+            var rewriter = state.GetRewriter(component);
+            Assert.AreEqual(expectedCode, rewriter.GetText());
+        }
+
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void LabelNotUsed_QuickFixWorks_MultipleLabels()
+        {
+            const string inputCode =
+@"Sub Foo()
+label1:
+dim var1 as variant
+label2:
+goto label1:
+End Sub";
+
+            const string expectedCode =
+@"Sub Foo()
+label1:
+dim var1 as variant
+
+goto label1:
+End Sub"; ;
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new LineLabelNotUsedInspection(state);
+            new RemoveUnusedDeclarationQuickFix(state).Fix(inspection.GetInspectionResults().First());
+
+            var rewriter = state.GetRewriter(component);
+            Assert.AreEqual(expectedCode, rewriter.GetText());
+        }
+
     }
 }
